+++ tests/testthat/test-collapseDegCreToGene.R
@@ -3,9 +3,9 @@ test_that("collapseDegCreToGene", {
   data("DexNR3C1")
   
   subDegGR <- 
-    DexNR3C1$DegGR[GenomicRanges::seqnames(DexNR3C1$DegGR)=="chr1"]
+    DexNR3C1$DegGR[GenomeInfoDb::seqnames(DexNR3C1$DegGR)=="chr1"]
   subCreGR <- 
-    DexNR3C1$CreGR[GenomicRanges::seqnames(DexNR3C1$CreGR)=="chr1"]
+    DexNR3C1$CreGR[GenomeInfoDb::seqnames(DexNR3C1$CreGR)=="chr1"]
   
   degCreResListDexNR3C1 <- DegCre::runDegCre(DegGR=subDegGR,
                                              DegP=subDegGR$pVal,
@@ -15,14 +15,13 @@ test_that("collapseDegCreToGene", {
                                              CreLfc=subCreGR$logFC,
                                              verbose=FALSE)
   
-  collapseDegCreResList <- collapseDegCreToGene(degCreResListDexNR3C1,
-                                                useParallel=FALSE)
+  collapseDegCreResList <- collapseDegCreToGene(degCreResListDexNR3C1)
   
   collapseHits  <- collapseDegCreResList$degCreHits
   
-  testIndices <- c(1,100,1000,5000,10000,15000,25000,34000)
+  testIndices <- c(1,100,1000,5000,10000,15000,17500,19989)
   
-  expectVals <- c(1,30,156,404,868,1573,2260,2775)
+  expectVals <- c(1,51,222,711,1651,2321,2508,2784)
   
   expect_equal(queryHits(collapseHits)[testIndices],expectVals)